Make build process create 'original' Elite separate from 'harmless'
The code is assembled with the `OPTION_ORIGINAL` symbol so that code can always provide the original code (via conditional ifs) if we change it. We will always build an original floppy disk image whenever we build Elite : Harmless, so as to ensure we never accidentally lose / alter the original code.

* Name and index most of the hulls. This will allow us to start using proper names where hulls are referenced, which should make it clearer what ship different pieces of code is working with

* Improvement to current memory map