Major rework of linker script
All-in on Exomizer; we now build seperate PRG files for disparate memory regions and rely on Exomizer to combine them into one payload and restore them to their intended location.

This has removed the undesired storing of the empty bitmap screen in the final PRG file as well as greatly simplifying some linker script complexity trying to deal with padding bytes between segments.

The run-time variable space at $0400..$0700 is now defined as a BSS segment.