Improvements to memory layout / linking
Given that the linker script is what decides the memory layout (particularly if the VIC bank is changed), we will need to produce a linker script for different memory layouts (such as VIC bank 1 or 3 and a cartridge layout).

Therefore it would be wiser, with a cartridge version planned and [possibly] moving the VIC bank, to define memory addresses ONLY within the linker script and export to the assembly source, rather than calculate them there (as we are currently doing).