Separate "DATA_SAVE" from the middle of "code_1D81.asm"
This extracts the "DATA_SAVE" segment from the middle of "code_1D81.asm", with the rest of the file broken out to "code_2A84.asm"

What we're doing here is trying to rebuild the way the game is assembled -- instead of lots of separate objects, we'll use a single master file that includes everything into one big object.
This should make assembling lots of slightly different versions of Elite much simpler.