First draft on the new assembling method
This changeset drastically reduces the number of separate compiler objects (and therefore imports/exports) by using a main file that includes all the files into one big super-object. Each majot style of game can use its own super-object (original / harmless), giving a more immediate and obvious source-code heirarchy.

This commit makes the necessary changes and gets the source code compiling again, but this does not represent the final outcome! More cleanup and documentation is required.

The goal is to reduce the number code segments used and to decouple the segment requirements of the original code from the Harmless builds.