Initial work on making a high-RAM (VIC bank #3) version
This builds a version of elite-harmless that uses VIC bank #3, i.e. $C000-$FFFF.
This places the bitmap under the KERNAL and allows for far greater contiguous memory, with a few caveats.
For one, the game is super crash prone right now, with some bugs regarding I/O interference.

Some rearranging of the screen initialisation has been done to fix conflict with I/O and to reduce grabage graphics on screen during initilisation.