Include the default bitmap image in the PRG
The original game does not include the bitmap screen data, instead constructing a screen programatically.

This commit now writes a default bitmap screen into the bitmap region in the PRG file.

This will allow us to do away with lots of code used to regularly reconstruct the screen (this code cleanup to follow)

NB: now that the default HUD is already in the PRG, we could do away with storing the copy and instead make a copy during initialisation.

The linker configs have been simplified to use a single PRG file, since we can rely upon Exomizer to compress the large amount of wasted space in the bitmap region that's now included.