Segmentise hulls
Instead of having the order of hull definitions given by the tables of data, we now append to each table for each hull. We can therefore now choose to reorder hulls or add / remove / replace hulls for different builds!

This commit also fixed the positions of the tables at D062->D063 & D083->D084 according to the way hulls are indexed.