Begin separating message IDs from tokens
When calling `print_docked_str` the text-database uses an ID 1-255, which is a separate concept from the bytes in those messages [tokens], which can self-reference via a separate set of IDs.

This commit changes a bunch of constant / label names to make this separation clearer, where by the text-database uses "msg" to indicate message-ID and "tkn" to indicate token-IDs.