DO YOU EVEN DECODE, BRO?
Used `.charmap` to redefine the ASCII chars in strings to the scrambled token values (for single characters), allowing us to do away with the "_A" symbols for single-letters (these are still defined, but go unused)

Filled in a bunch more of the planet description expansions.

Identified the docked token function that prints the target-system name, with an -"ian" suffix, e.g. "Lavian"