More understanding
* Created a "heap" memory zone:

The game relies upon an empty block of RAM to allocate memory top-down for the list of lines to draw for each ship in play. The bottom of the heap contains the instance-data for the ships.

In the original game, this heap ran from $F900 fo the ship storage, and $FFC0 down for the lines.

I've now assigned this zone in the linker config, removing one of the last hard-coded segment addresses in the game.

* Started documenting the main game loop

* Identified some colour-nybble literals related to the missile indicator HUD