Optimise polyobj<->zero-page copying
For every ship being processed, the ship instance data is copied to the zero-page, modified, and copied back.
With a number of ships in the area this can add up to a lot of cycles.
This commit provides optimised copy routines, depending on build option.
The code has been moved to "vars_polyobj.asm" as a pair of macros so that they can be inserted into the right place in the code.