Change segmentation
I've reintroduced the code segments used in the original, for the harmless builds, so that we can split the code into more segments without having to use separate files.
The plan is to reduce the number of macros that insert chunks of code into specific places, and use segments instead.